Archivo env.example

+++ src/controllers/appointmentsController.js
@@ -1,11 +1,16 @@
 const { where } = require("sequelize");
 const appointmentModel = require("../models/appointmentModel")
+const jwt = require('jsonwebtoken');
+
 //Crear el modelo de appointments
 
 const getAppointments = async (req,res) =>{
     try {
-        const  { page = 1, limit = 10, sort = "ASC", status } = req.query;
+        const token = req.headers.authorization?.split(" ")[1];
+        const decoded = jwt.verify(token, process.env.SECRET_KEY);  // Asegúrate de que el secreto es correcto
+        const { id: userId, role } = decoded;  // Extraemos id y role del payload del token
 
+        const  { page = 1, limit = 10, sort = "ASC", status } = req.query;
         const pageInt = parseInt(page)
         const limitInt = parseInt(limit)
         const offset = (pageInt - 1) * limitInt
@@ -14,9 +19,11 @@ const getAppointments = async (req,res) =>{
         if (status){
             filter.status = status
         }
-
         const order = [["createdAt", sort.toUpperCase() === "DESC" ? "DESC" : "ASC"]];
 
+        if(role == "user"){
+            filter.userId = userId
+        }
         const appointments = await appointmentModel.findAll({
             where: filter, // Aquí pasamos los filtros dinámicos
             limit: limitInt,  // Límite de registros por página
@@ -31,6 +38,17 @@ const getAppointments = async (req,res) =>{
 
 const postAppointments = async(req,res)=>{
     try {
+        const { date, time } = req.body;
+        const existingAppointment = await appointmentModel.findOne({
+            where: {
+                date, 
+                time 
+            }
+        });
+        if (existingAppointment) {
+            return res.status(400).json({ message: "Ya existe un turno en este horario." });
+        }
+
         await appointmentModel.create(req.body)
         res.status(201).json("Registro exitoso")
     } catch (error) {
@@ -40,6 +58,22 @@ const postAppointments = async(req,res)=>{
 
 const deleteAppointments = async(req,res)=>{
     try {
+        const token = req.headers.authorization?.split(" ")[1];
+        const decoded = jwt.verify(token, process.env.SECRET_KEY);  
+        const { id: userId, role } = decoded;
+        // Buscar el turno por ID
+        const appointment = await appointmentModel.findOne({
+            where: { id: req.params.id }
+        });
+
+        if (!appointment) {
+            return res.status(404).json({ message: "Turno no encontrado" });
+        }
+        // Si el rol es 'user', validamos que el turno pertenezca a este usuario
+        if (role === "user" && appointment.userId !== userId) {
+            return res.status(403).json({ message: "No tienes permiso para eliminar este turno" });
+        }
+
         await appointmentModel.destroy({
             where:{id:req.params.id}
         })
@@ -51,8 +85,20 @@ const deleteAppointments = async(req,res)=>{
 
 const getAppointmentId =async (req,res)=>{
     try {
-        const appointment = await appointmentModel.findByPk(req.params.id)
-        res.status(200).json(appointment)
+        const token = req.headers.authorization?.split(" ")[1];
+        const decoded = jwt.verify(token, process.env.SECRET_KEY);  
+        const { id: userId, role } = decoded;
+        const appointment = await appointmentModel.findByPk(req.params.id);
+        if (!appointment) {
+            return res.status(404).json({ message: "Turno no encontrado" });
+        }
+        // Si el rol es 'user', validamos que el turno pertenezca a este usuario
+        if (role === "user" && appointment.userId !== userId) {
+            return res.status(403).json({ message: "No tienes permiso para ver este turno" });
+        }
+        // Si pasa las validaciones, mostramos el turno
+        res.status(200).json(appointment);
+
     } catch (error) {
         res.status(500).json({message:error.message})
     }
@@ -60,6 +106,17 @@ const getAppointmentId =async (req,res)=>{
 
 const putAppointmentsId = async (req,res)=>{
     try {
+        const token = req.headers.authorization?.split(" ")[1];
+        const decoded = jwt.verify(token, process.env.SECRET_KEY);  
+        const { id: userId, role } = decoded;
+        const appointment = await appointmentModel.findByPk(req.params.id);
+        if (!appointment) {
+            return res.status(404).json({ message: "Turno no encontrado" });
+        }
+        // Si el rol es 'user', validamos que el turno pertenezca a este usuario
+        if (role === "user" && appointment.userId !== userId) {
+            return res.status(403).json({ message: "No tienes permiso para ver este turno" });
+        }
         await appointmentModel.update(req.body, {
             where:{id:req.params.id}
         })
